requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.5.2
async-timeout==4.0.2
attrs==21.4.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.4
defusedxml==0.7.1
Django==4.0.6
django-allauth==0.51.0
django-ckeditor==6.4.2
django-crispy==1.8.1
django-js-asset==2.0.0
django-rosetta==0.9.8
django-social-share==2.2.1
frozenlist==1.3.0
idna==3.3
multidict==6.0.2
oauthlib==3.2.0
Pillow==9.2.0
polib==1.1.1
pycparser==2.21
PyJWT==2.4.0
python3-openid==3.2.0
requests==2.28.1
requests-oauthlib==1.3.1
sqlparse==0.4.2
telepot==12.7
tzdata==2022.1
urllib3==1.26.10
yarl==1.7.2
